[Fix] 씬 전환 후 광고 로드 쇼 기능 안되는 문제 해결
