+++ Assets/Personal work/JGH/Scene/JGH_OutGameUI.unity
@@ -5373,6 +5373,50 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 579310572}
   m_CullTransparentMesh: 1
+--- !u!1 &588997585
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 588997587}
+  - component: {fileID: 588997586}
+  m_Layer: 0
+  m_Name: AdSystem
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &588997586
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 588997585}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2627e49738aa410a916e38ef99be9401, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!4 &588997587
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 588997585}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 451.79474, y: 1159.9288, z: -5.120954}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &598939389
 GameObject:
   m_ObjectHideFlags: 0
@@ -9911,6 +9955,7 @@ GameObject:
   - component: {fileID: 1145936667}
   - component: {fileID: 1145936666}
   - component: {fileID: 1145936668}
+  - component: {fileID: 1145936669}
   m_Layer: 5
   m_Name: Bmx (2)
   m_TagString: Untagged
@@ -10018,19 +10063,20 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 1145936666}
   m_OnClick:
     m_PersistentCalls:
-      m_Calls:
-      - m_Target: {fileID: 1003951956}
-        m_TargetAssemblyTypeName: OutGameUiController, Assembly-CSharp
-        m_MethodName: ShowAD
-        m_Mode: 1
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
+      m_Calls: []
+--- !u!114 &1145936669
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1145936664}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e2de314c90ad41ce8b797ce2d5e120b0, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  bmxButton: {fileID: 1145936668}
 --- !u!1 &1150757861
 GameObject:
   m_ObjectHideFlags: 0
@@ -13126,6 +13172,7 @@ GameObject:
   - component: {fileID: 1563603220}
   - component: {fileID: 1563603219}
   - component: {fileID: 1563603221}
+  - component: {fileID: 1563603222}
   m_Layer: 5
   m_Name: Bmx
   m_TagString: Untagged
@@ -13233,19 +13280,20 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 1563603219}
   m_OnClick:
     m_PersistentCalls:
-      m_Calls:
-      - m_Target: {fileID: 1003951956}
-        m_TargetAssemblyTypeName: OutGameUiController, Assembly-CSharp
-        m_MethodName: LoadAD
-        m_Mode: 1
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
+      m_Calls: []
+--- !u!114 &1563603222
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1563603217}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c9eb0e9f69f4c6e4688b9f390cc855a9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  bmxButton: {fileID: 1563603221}
 --- !u!1 &1567448860
 GameObject:
   m_ObjectHideFlags: 0
@@ -18508,9 +18556,9 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _bgmClips:
   - name: InGameMusic
-    audioClip: {fileID: 8300000, guid: db56b5a64e06a8048a4ec83c4ca3bd6a, type: 3}
+    audioClip: {fileID: 8300000, guid: 3a13ea34c92f59f4dad171c7d306c42f, type: 3}
   - name: OutGameMusic
-    audioClip: {fileID: 8300000, guid: a328ad3c6eb967d4e8ec388e0740db1b, type: 3}
+    audioClip: {fileID: 8300000, guid: f3ab59330f8892545b08cd9b7fdc8a02, type: 3}
   _sfxClips:
   - name: GameInSfx
     audioClip: {fileID: 8300000, guid: 6d2c92f0182a7964c918b09845029f19, type: 3}
@@ -18615,6 +18663,7 @@ SceneRoots:
   - {fileID: 1916669641}
   - {fileID: 945860154}
   - {fileID: 1633761895}
+  - {fileID: 588997587}
   - {fileID: 1003951957}
   - {fileID: 9202291117675283618}
   - {fileID: 3467843919896283239}
