+++ Assets/Personal work/JGH/Scripts/FuntionButtons/CoinAdLoadButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c9eb0e9f69f4c6e4688b9f390cc855a9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
