+++ Assets/Personal work/JGH/Scripts/Setting/AdSystem.cs
@@ -1,0 +1,98 @@
+using GoogleMobileAds.Api;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class AdSystem : MonoBehaviour
+{
+    public static AdSystem Instance { get; private set; }
+    
+    private RewardedInterstitialAd _rewardedInterstitialAd;
+    
+    protected void Awake()
+    {
+        if (Instance != null && Instance != this)
+        {
+            Destroy(gameObject);
+            return;
+        }
+        Instance = this;
+        DontDestroyOnLoad(gameObject);
+    }
+
+    private void Start()
+    {
+        MobileAds.Initialize((InitializationStatus initstatus) =>
+        {
+            if (initstatus == null)
+            {
+                Debug.LogError("Google Mobile Ads initialization failed.");
+                return;
+            }
+
+            LoadAD();
+            Debug.Log("Google Mobile Ads initialization complete.");
+        }); 
+    }
+
+    /// <summary>
+    /// Loads the rewarded interstitial ad.
+    /// </summary>
+    public void LoadAD()
+    {
+        // Clean up the old ad before loading a new one.
+        if (_rewardedInterstitialAd != null)
+        {
+            _rewardedInterstitialAd.Destroy();
+            _rewardedInterstitialAd = null;
+        }
+
+        Debug.Log("Loading the rewarded interstitial ad.");
+
+        // create our request used to load the ad.
+        var adRequest = new AdRequest();
+        adRequest.Keywords.Add("unity-admob-sample");
+
+        // send the request to load the ad.
+        RewardedInterstitialAd.Load("ca-app-pub-3940256099942544/5354046379", adRequest,
+        (RewardedInterstitialAd ad, LoadAdError error) =>
+        {
+            // if error is not null, the load request failed.
+            if (error != null || ad == null)
+            {
+                Debug.LogError("rewarded interstitial ad failed to load an ad " +
+                               "with error : " + error);
+                return;
+            }
+
+            Debug.Log("Rewarded interstitial ad loaded with response : "
+                      + ad.GetResponseInfo());
+
+            _rewardedInterstitialAd = ad;
+            
+            
+            // 이벤트 등록
+            _rewardedInterstitialAd.OnAdFullScreenContentClosed += () =>
+            {
+                Debug.Log("Ad closed → reload next ad");
+                LoadAD();
+            };
+
+            _rewardedInterstitialAd.OnAdFullScreenContentFailed += (AdError adError) =>
+            {
+                Debug.LogError("Ad failed to show: " + adError);
+            };
+        });
+        
+    }
+    public void ShowAD()
+    {
+        if (_rewardedInterstitialAd != null && _rewardedInterstitialAd.CanShowAd())
+        {
+            _rewardedInterstitialAd.Show((Reward reward) =>
+            {
+                // TODO: Reward the user.
+                Debug.Log($"User earned reward: {reward.Amount} {reward.Type}");
+            });
+        }
+    }
+}
